// js/dom_types.d.ts
type BlobPart = BufferSource | Blob | string;
type FormDataEntryValue = File | string;
declare type EventListenerOrEventListenerObject =
  | EventListener
  | EventListenerObject;

interface Element {
  // TODO
}

interface HTMLFormElement {
  // TODO
}

type EndingType = "tranparent" | "native";

interface BlobPropertyBag {
  type?: string;
  ending?: EndingType;
}
// ...
export interface Blob {
  readonly size: number;
  readonly type: string;
  slice(start?: number, end?: number, contentType?: string): Blob;
}